feat: Refactor account tab widget and update transfer amount type to double

+++ lib/features/transfer/data/model/request/transfer_request.dart
@@ -1,21 +1,20 @@
-
 class TransferRequest {
   String receiverEmail;
   bool usebalance;
-  int amount;
+  double amount;
   String? cardId;
 
   TransferRequest({
     required this.receiverEmail,
-      required this.usebalance,
+    required this.usebalance,
     required this.amount,
     this.cardId,
   });
 
   Map<String, dynamic> toJson() => {
-        'receiverEmail': receiverEmail,
-        'amount': amount,
-        'usebalance': usebalance,
-        'cardId': cardId,
-      };
+    'receiverEmail': receiverEmail,
+    'amount': amount,
+    'usebalance': usebalance,
+    'cardId': cardId,
+  };
 }
